+++ src/main/java/com/example/project_pulse_backend/dto/response/CreateAccountResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,5 +15,5 @@ public class CreateAccountResponse {
     private String email;
     private String userName;
     private String position;
-    private Float resourceRate;
+    private BigDecimal resourceRate;
 }
